refactor(plugin)!: relocate host.v1 servers to hostcap pkg behind the port (holomush-eykuh.2)
