+++ internal/plugin/hostcap/servers.go
@@ -1,11 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2026 HoloMUSH Contributors
 
-package goplugin
+package hostcap
 
 import (
 	"context"
 	"log/slog"
+	"math"
 	"time"
 
 	"github.com/oklog/ulid/v2"
@@ -17,7 +18,7 @@ import (
 	"github.com/holomush/holomush/internal/access/policy/types"
 	"github.com/holomush/holomush/internal/core"
 	"github.com/holomush/holomush/internal/eventbus/cursor"
-	"github.com/holomush/holomush/internal/plugin/hostcap"
+	plugins "github.com/holomush/holomush/internal/plugin"
 	"github.com/holomush/holomush/internal/plugin/pluginauthz"
 	"github.com/holomush/holomush/internal/session"
 	"github.com/holomush/holomush/internal/settings"
@@ -29,9 +30,10 @@ import (
 // services. After the holomush-eykuh.1 decomposition (Task 12) these are the
 // SINGLE source of the authenticated host-callback logic: the former monolithic
 // pluginHostServiceServer is gone and each authoritative handler body lives on
-// the per-capability server for its domain (one capability per service). The
-// host internals (*Host, session.FocusKey, settings stores) are reached
-// directly; there is no longer a pluginv1<->hostv1 translation hop.
+// the per-capability server for its domain (one capability per service). After
+// the holomush-eykuh.2 relocation these bodies live in the runtime-neutral
+// hostcap package so both the binary (goplugin) and Lua runtimes consume the
+// SAME server implementations through the HostCapabilities port (INV-PLUGIN-49).
 //
 // hostCapabilityBase carries the host handle and the mTLS-bound plugin name; every
 // per-capability server embeds it so the shared host wiring is declared once.
@@ -43,8 +45,23 @@ import (
 // the host mutex — every mutable-state read is a port method that locks
 // internally.
 type hostCapabilityBase struct {
-	host       hostcap.HostCapabilities
+	host       HostCapabilities
 	pluginName string
+}
+
+// NewBase builds a hostCapabilityBase binding the given HostCapabilities port
+// adapter and the mTLS-bound (binary) or wiring-time-established (Lua) plugin
+// name. The base is embedded into every per-capability server. hostCapabilityBase
+// stays unexported — NewBase is the only construction surface — so callers wire
+// servers through NewBase + RegisterCapabilities (or the per-server constructors)
+// rather than reaching into the struct.
+//nolint:revive // unexported-return is intentional: hostCapabilityBase stays
+// unexported by design (the public construction surface is NewBase +
+// RegisterCapabilities). Callers only ever pass the returned value straight back
+// into RegisterCapabilities / the New*Server constructors, never reaching into
+// the struct — so the opaque return type is the intended ergonomics, not a wart.
+func NewBase(host HostCapabilities, pluginName string) hostCapabilityBase {
+	return hostCapabilityBase{host: host, pluginName: pluginName}
 }
 
 // --- focusServer (FocusService: 8 RPCs) -------------------------------------
@@ -992,4 +1009,181 @@ func (s *commandRegistryServer) GetCommandHelp(ctx context.Context, req *hostv1.
 type kvServer struct {
 	hostv1.UnimplementedKVServiceServer
 	hostCapabilityBase
+}
+
+// --- converters & helpers ---------------------------------------------------
+//
+// These conversion helpers moved with the server bodies (holomush-eykuh.2): they
+// are the proto↔domain glue the relocated servers call and have no consumer
+// outside this package.
+
+// leaveByTargetResultToProto converts the host-side sweep result to the
+// wire format. Callers reconstruct partial-success state from
+// succeeded + len(failed_session_ids) == total_scanned.
+func leaveByTargetResultToProto(r session.LeaveByTargetResult) *hostv1.LeaveFocusByTargetResponse {
+	resp := &hostv1.LeaveFocusByTargetResponse{
+		Succeeded:    clampCountToInt32(r.Succeeded),
+		TotalScanned: clampCountToInt32(r.TotalScanned),
+	}
+	if len(r.Failed) > 0 {
+		resp.FailedSessionIds = make([]string, 0, len(r.Failed))
+		for _, f := range r.Failed {
+			resp.FailedSessionIds = append(resp.FailedSessionIds, f.SessionID)
+		}
+	}
+	return resp
+}
+
+// clampCountToInt32 narrows a Go int to proto int32 safely. The session count
+// is bounded by live-session capacity (far below math.MaxInt32 in any realistic
+// deployment), but explicit bounds keep gosec quiet and guard against future
+// 64-bit-only callers.
+func clampCountToInt32(n int) int32 {
+	switch {
+	case n < 0:
+		return 0
+	case n > math.MaxInt32:
+		return math.MaxInt32
+	default:
+		return int32(n)
+	}
+}
+
+// focusKeyToProto converts a session.FocusKey to the host.v1 FocusKey type.
+func focusKeyToProto(fk session.FocusKey) *hostv1.FocusKey {
+	return &hostv1.FocusKey{
+		Kind:     focusKindToProto(fk.Kind),
+		TargetId: fk.TargetID.String(),
+	}
+}
+
+// focusKindToProto maps session.FocusKind to host.v1 FocusKind.
+func focusKindToProto(k session.FocusKind) hostv1.FocusKind {
+	switch k {
+	case session.FocusKindScene:
+		return hostv1.FocusKind_FOCUS_KIND_SCENE
+	default:
+		return hostv1.FocusKind_FOCUS_KIND_UNSPECIFIED
+	}
+}
+
+// bytesToULID converts a 16-byte proto bytes field to ulid.ULID.
+// Returns INVALID_ULID error on wrong length (proto3 bytes ULID fields
+// carry the 16-byte binary form, not the 26-char string encoding).
+func bytesToULID(b []byte) (ulid.ULID, error) {
+	if len(b) != 16 {
+		return ulid.ULID{}, oops.Code("INVALID_ULID").
+			Errorf("expected 16-byte ULID, got %d bytes", len(b))
+	}
+	var id ulid.ULID
+	copy(id[:], b)
+	return id, nil
+}
+
+// autoFocusFailureReasonToProto maps the string reason from AutoFocusOnJoin
+// to the host.v1 FocusFailureReason enum.
+func autoFocusFailureReasonToProto(reason string) hostv1.FocusFailureReason {
+	switch reason {
+	case "membership_absent":
+		return hostv1.FocusFailureReason_FOCUS_FAILURE_REASON_MEMBERSHIP_ABSENT
+	case "connection_not_found":
+		return hostv1.FocusFailureReason_FOCUS_FAILURE_REASON_CONNECTION_NOT_FOUND
+	default:
+		return hostv1.FocusFailureReason_FOCUS_FAILURE_REASON_UNSPECIFIED
+	}
+}
+
+// protoToFocusKey converts a host.v1 FocusKey to the session.FocusKey domain type.
+func protoToFocusKey(pk *hostv1.FocusKey) (session.FocusKey, error) {
+	if pk == nil {
+		return session.FocusKey{}, oops.Code("INVALID_ARGUMENT").
+			Errorf("focus key is required")
+	}
+
+	targetID, err := ulid.Parse(pk.GetTargetId())
+	if err != nil {
+		return session.FocusKey{}, oops.Code("INVALID_ARGUMENT").
+			With("target_id", pk.GetTargetId()).
+			Wrap(err)
+	}
+
+	kind, err := protoToFocusKind(pk.GetKind())
+	if err != nil {
+		return session.FocusKey{}, err
+	}
+
+	return session.FocusKey{Kind: kind, TargetID: targetID}, nil
+}
+
+// protoToFocusKind maps host.v1 FocusKind to session.FocusKind.
+func protoToFocusKind(pk hostv1.FocusKind) (session.FocusKind, error) {
+	switch pk {
+	case hostv1.FocusKind_FOCUS_KIND_SCENE:
+		return session.FocusKindScene, nil
+	default:
+		return "", oops.Code("FOCUS_KIND_UNREGISTERED").
+			With("kind", pk.String()).
+			Errorf("unsupported focus kind: %s", pk.String())
+	}
+}
+
+const maxQueryStreamHistoryCount = 500
+
+// encodeHostEventCursor encodes an event ULID into an opaque host cursor
+// token for the plugin → host boundary. Seq is not available here (the
+// plugins.HistoryReader.ReplayTail interface returns core.Event without Seq),
+// so Seq=0 is used. The cold tier handles Seq=0 as "ID-only" fallback.
+// Returns nil on encoding failure (non-fatal; client cannot paginate from
+// this event but the page result is still valid).
+func encodeHostEventCursor(id ulid.ULID) []byte {
+	b, err := cursor.Encode(cursor.Cursor{
+		Version: cursor.CurrentVersion,
+		Epoch:   cursor.CurrentEpoch(),
+		Owner:   cursor.Owner{Kind: cursor.OwnerHost},
+		Host:    &cursor.HostCursor{Seq: 0, ID: id},
+	})
+	if err != nil {
+		return nil
+	}
+	return b
+}
+
+// coreEventToProto converts a core.Event to the host.v1 Event.
+func coreEventToProto(e core.Event) *hostv1.Event {
+	return &hostv1.Event{
+		Id:        e.ID.String(),
+		Stream:    e.Stream,
+		Type:      string(e.Type),
+		Timestamp: e.Timestamp.UnixMilli(),
+		ActorKind: e.Actor.Kind.String(),
+		ActorId:   e.Actor.ID,
+		Payload:   string(e.Payload),
+	}
+}
+
+// stampPluginActor resolves a plugin name to a core.Actor with a ULID-string
+// ID via the IdentityRegistry. Returns PLUGIN_UNREGISTERED_INVOKE if the
+// plugin is not active in the registry, or if the registry is nil (which is
+// operationally equivalent: "no registry" and "registry doesn't have plugin"
+// both mean the ULID cannot be resolved). This defensive nil-check keeps
+// existing test fixtures that construct a host directly without registering it
+// safe — they'll receive a clean error rather than a nil-pointer panic.
+//
+// A goplugin-private twin (goplugin.stampPluginActor) backs the binary host's
+// own self-token paths; this copy is the runtime-neutral one the relocated
+// emitServer.RequestEmitToken calls through the port snapshot. Both operate
+// purely on the plugins.IdentityRegistry port, so they cannot drift in policy.
+func stampPluginActor(reg plugins.IdentityRegistry, name string) (core.Actor, error) {
+	if reg == nil {
+		return core.Actor{}, oops.Code("PLUGIN_UNREGISTERED_INVOKE").
+			With("plugin", name).
+			Errorf("IdentityRegistry not configured on Host")
+	}
+	id, ok := reg.IDByName(name)
+	if !ok {
+		return core.Actor{}, oops.Code("PLUGIN_UNREGISTERED_INVOKE").
+			With("plugin", name).
+			Errorf("plugin not registered in IdentityRegistry")
+	}
+	return core.Actor{Kind: core.ActorPlugin, ID: id.String()}, nil
 }
